Modularise the code
